move common constants to constants.py
BusType, ModeName
