+++ wakepy/core/constants.py
@@ -10,6 +10,11 @@ class SystemName(StrEnum):
     DARWIN = "darwin"
 
 
+class ModeName(StrEnum):
+    KEEP_RUNNING = "keep.running"
+    KEEP_PRESENTING = "keep.presenting"
+
+
 class ControlMsg(StrEnum):
     """Send to worker threads"""
 
@@ -21,3 +26,10 @@ class WorkerThreadMsgType(StrEnum):
 
     OK = auto()
     EXCEPTION = auto()
+
+
+class BusType(StrEnum):
+    """Type of D-Bus bus."""
+
+    SESSION = auto()
+    SYSTEM = auto()
