+++ wakepy/core/mode.py
@@ -5,7 +5,7 @@ from abc import ABC
 
 from .activationmanager import ModeActivationManager
 from .activationresult import ActivationResult
-from .constant import StringConstant
+from .constants import ModeName
 
 if typing.TYPE_CHECKING:
     from types import TracebackType
@@ -13,11 +13,6 @@ if typing.TYPE_CHECKING:
 
     from .dbus import DbusAdapter, DbusAdapterTypeSeq
     from .method import Method
-
-
-class ModeName(StringConstant):
-    KEEP_RUNNING = "keep.running"
-    KEEP_PRESENTING = "keep.presenting"
 
 
 class ModeExit(Exception):
